+++ models/incomit.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Date
+from sqlalchemy import Column, Integer, Date, Numeric
 from db_connect import Base
 
 
@@ -6,5 +6,5 @@ class Incomes(Base):
     __tablename__ = "incomes"
     id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(Integer, nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(Numeric, nullable=False)
     date_receipt = Column(Date, nullable=False)
