+++ models/planshet.py
@@ -1,5 +1,5 @@
 from db_connect import Base
-from sqlalchemy import String, Column, Integer, Float
+from sqlalchemy import String, Column, Integer, Numeric
 
 
 class Planshets(Base):
@@ -7,7 +7,7 @@ class Planshets(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(255), nullable=False)
     category_id = Column(Integer, nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(Numeric, nullable=False)
     brand = Column(String(255), nullable=False)
     screen_type = Column(String(255), nullable=False)
     camera = Column(Integer, nullable=False)
@@ -17,9 +17,9 @@ class Planshets(Base):
     ram_size = Column(Integer, nullable=False)
     rom_size = Column(Integer, nullable=False)
     year = Column(Integer, nullable=False)
-    display = Column(Float, nullable=False)
-    weight = Column(Float, nullable=False)
+    display = Column(Numeric, nullable=False)
+    weight = Column(Numeric, nullable=False)
     discount = Column(Integer, nullable=False)
-    discount_price = Column(Integer, nullable=False)
+    discount_price = Column(Numeric, nullable=False)
     count = Column(Integer, nullable=False)
     discount_time = Column(Integer)
