+++ models/telephone.py
@@ -1,5 +1,5 @@
 from db_connect import Base
-from sqlalchemy import String, Column, Integer, Float
+from sqlalchemy import String, Column, Integer, Numeric
 
 
 class Telephones(Base):
@@ -7,7 +7,7 @@ class Telephones(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(255), nullable=False)
     category_id = Column(Integer)
-    price = Column(Integer)
+    price = Column(Numeric)
     brand = Column(String(255), nullable=False)
     model = Column(String(255), nullable=False)
     color = Column(String(255), nullable=False)
@@ -15,12 +15,12 @@ class Telephones(Base):
     ram_size = Column(Integer, nullable=False)
     rom_size = Column(Integer, nullable=False)
     year = Column(Integer, nullable=False)
-    display = Column(Float, nullable=False)
-    weight = Column(Float, nullable=False)
+    display = Column(Numeric, nullable=False)
+    weight = Column(Numeric, nullable=False)
     battery = Column(Integer, nullable=False)
     camera = Column(Integer, nullable=False)
     self_camera = Column(Integer, nullable=False)
     discount = Column(Integer, nullable=False)
-    discount_price = Column(Integer, nullable=False)
+    discount_price = Column(Numeric, nullable=False)
     count = Column(Integer, nullable=False)
     discount_time = Column(Integer, nullable=False)
